Add files via upload
The goal of this study is to predict the presence of heart disease using the Cleveland database, which includes 76 attributes. Past research has focused on a subset of 14 key attributes, including age, sex, and various medical predictors. For this study, the 14 variables are used as features for training the KNN classifier.  More detailed information about the dataset can be found on the Kaggle website:https://www.kaggle.com/datasets/redwankarimsony/heart-disease-data. Additionally, I do not directly use of functions from the scikit-learn package for this study.